+++ spec/integrations/pro/consumption/periodic_jobs/during_error_with_recovery_spec.rb
@@ -6,8 +6,8 @@
 # If we explicitely want to tick during recovery, it should be possible
 
 setup_karafka(allow_errors: true) do |config|
-  config.pause_timeout = 1_000
-  config.pause_max_timeout = 1_500
+  config.pause.timeout = 1_000
+  config.pause.max_timeout = 1_500
   config.max_wait_time = 100
   config.max_messages = 1
 end