+++ spec/integrations_helper.rb
@@ -74,9 +74,9 @@ def setup_karafka(
     config.client_id = DT.consumer_group
     # Prevents conflicts when running in parallel
     config.group_id = DT.consumer_group
-    config.pause_timeout = 1
-    config.pause_max_timeout = 1
-    config.pause_with_exponential_backoff = false
+    config.pause.timeout = 1
+    config.pause.max_timeout = 1
+    config.pause.with_exponential_backoff = false
     config.max_wait_time = 500
     config.shutdown_timeout = 30_000
     config.swarm.nodes = 2